requirements.txt
aiokafka==0.2.2
kafka-python==1.3.1
nltk==3.2.2
numpy==1.12.1
oauthlib==2.0.2
requests==2.13.0
requests-oauthlib==0.8.0
Shapely==1.5.17.post1
six==1.10.0
tweepy==3.5.0
websockets==3.3
